Separating test class for arm state
